+++ network/src/main/java/com/github/adrian99/neuralnetwork/learning/error/ErrorFunction.java
@@ -1,6 +1,6 @@
 package com.github.adrian99.neuralnetwork.learning.error;
 
 public interface ErrorFunction {
-    double apply(double[] networkOutputs, double[] expectedOutputs);
-    double applyDerivative(double networkOutput, double targetOutput);
+    double apply(double[] networkOutputs, int[] expectedOutputs);
+    double applyDerivative(double networkOutput, int targetOutput);
 }
